+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Add AsgAmiId parameter to support self-service marketplace updates
+
 # 4.1.9
 
 * Fix metadata for new parameters